requirements.txt
pillow
opencv-python
torch
torchvision
torchaudio
tqdm
numpy